Quick tweaking and additional checks

+++ native/jobs/cache/clear.php
@@ -7,6 +7,8 @@
     class Clear extends Job {
         public function run(?array $context): ?string
         {
+            if(!file_exists(cache_path($context['route']))) return null;
+
             unlink(cache_path($context['route']));
 
             return null;
